refactor identity prop type in repository an controller

+++ SMSInteraction.Repository/Interfaces/IAnswerRepository.cs
@@ -6,8 +6,8 @@ namespace SMSInteraction.Repository.Interfaces;
 
 public interface IAnswerRepository : IGenericRepository<Answer>
 {
-    List<AnswerResultDto> GetAnswersOfInteraction(int interactionId);
-    AnswerResultDto? GetAnswer(int answerId);
-    void Add(int interactionId, AnswerAddDto dto);
-    void EditAnswer(int answerId, AnswerEditDto dto);
+    List<AnswerResultDto> GetAnswersOfInteraction(long interactionId);
+    AnswerResultDto? GetAnswer(long answerId);
+    void Add(long interactionId, AnswerAddDto dto);
+    void EditAnswer(long answerId, AnswerEditDto dto);
 }
